[Bug]: `#maLazyFrom:` Infinite Loop
Old implementation: In the *getter*, use the description to access the current value, which for accessor selectors uses the *getter*... uh oh
New Implementation: Try to directly read an instance variable with the same name as the read selector. NB this is not robust because it only works with a selector accessor with matching inst var name, but is "less-wrong" than the previous implementation, and this is currently a convenience feature that is only used by 1 person (me). A note is in the method comment to this effect.