[Enh]: Descriptions - Real Names
It can often be useful for a description to have a canonical name. For some accessors there is an obvious default e.g. read selector or variable name.

Descriptions *did* have a "name", but it was really a string explaining the type of the description itself e.g. "1:M..." for to-many descriptions. 

Here, we rename the existing `#name` behavior to `#type` and reimplement `#name` as a configurable canonical specifier.