[Bug]: Chain Accessors Fields Flipped
Per the class comment: "To read and write a value the ==accessor== is performed on the given model and the result is passed into the ==next== accessor."
However, the implementation was using the ==next== accessor *first* instead of last.

Having trouble committing the fixes to the tests so there will likely be at least one more commit for this.